constructors

+++ src/Camouflage.h
@@ -6,10 +6,10 @@
 
 class Camouflage : public Decorator {
     private:
-        float hidingCoeff;
+        double hidingCoeff;
     
     public:
-        Camouflage(Bestiole &b, float hidingCoeff);
+        Camouflage(Bestiole &b, double hidingCoeff);
         void draw(UImg &support, double x, double y, double orientation) override;
 };
